chore: extract create credential into separate module
