+++ crates/bitwarden/src/platform/fido2/mod.rs
@@ -1,8 +1,11 @@
+pub mod client_create_credential;
 mod fido2;
 mod user_interface;
 
-pub(crate) use fido2::{client_create_credential, client_get_assertion};
-pub use fido2::{Fido2ClientCreateCredentialRequest, Fido2ClientGetAssertionRequest};
+pub use client_create_credential::Fido2ClientCreateCredentialRequest;
+pub(crate) use fido2::client_get_assertion;
+
+pub use fido2::Fido2ClientGetAssertionRequest;
 pub use user_interface::{
     Fido2GetAssertionUserInterface, Fido2MakeCredentialUserInterface, NewCredentialParams,
     NewCredentialResult, VaultItem,
